AdventOfCode2023/Tests/Day05Tests.cs
using AdventOfCode2023.Solutions;

namespace AdventOfCode2023.Tests
{
    [TestClass]
    public class Day05Tests
    {
        private readonly string _sample =
@"seeds: 79 14 55 13

seed-to-soil map:
50 98 2
52 50 48

soil-to-fertilizer map:
0 15 37
37 52 2
39 0 15

fertilizer-to-water map:
49 53 8
0 11 42
42 0 7
57 7 4

water-to-light map:
88 18 7
18 25 70

light-to-temperature map:
45 77 23
81 45 19
68 64 13

temperature-to-humidity map:
0 69 1
1 0 69

humidity-to-location map:
60 56 37
56 93 4";

        private readonly string _input = File.ReadAllText("Input/Day05.txt");

        private Day05 _day = new();

        [TestInitialize]
        public void Initialize()
        {
            _day = new Day05();
        }

        [TestMethod]
        public void Part1Sample()
        {
            Assert.AreEqual(35, _day.Part1(_sample));
        }

        [TestMethod]
        public void Part1()
        {
            Assert.AreEqual(486613012, _day.Part1(_input));
        }

        [TestMethod]
        public void Part2Sample()
        {
            Assert.AreEqual(46, _day.Part2(_sample));
        }

        [TestMethod]
        public void Part2()
        {
            Assert.AreEqual(56931769, _day.Part2(_input));
        }
    }
}